requirements.txt
Flask==3.0.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
flask-mysql==1.5.2
pymysql
fontawesomefree